backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.128.0",
    "faster-whisper>=1.2.1",
    "groq>=1.0.0",
    "httpx>=0.28.1",
    "mcp[cli]>=1.25.0",
    "numpy>=2.4.0",
    "onnxruntime>=1.23.2",
    "openwakeword>=0.4.0",
    "piper-tts>=1.3.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.21",
    "requests>=2.32.5",
    "soundfile>=0.13.1",
    "spotipy>=2.25.2",
    "torch>=2.9.1",
    "torchaudio>=2.9.1",
    "uvicorn[standard]>=0.40.0",
    "vadersentiment>=3.3.2",
]
